+++ it/server/src/test/resources/com/walmartlabs/concord/it/server/dynamicTask/tasks/test.groovy
@@ -1,0 +1,11 @@
+import com.walmartlabs.concord.sdk.Task
+
+import javax.inject.Named
+
+@Named("myTask")
+class MyTask implements Task {
+
+    void hey(String name) {
+        println "Hey, ${name}"
+    }
+}